Update issue 3024 Making sure visibility uses the right heights from the DEM.
Former-commit-id: 12e43b7af32ae08746dbe01d081220646e40e625

+++ invest_natcap/scenic_quality/scenic_quality_cython_core.pyx
@@ -628,7 +628,7 @@ def sweep_through_angles( \
     np.ndarray[np.float64_t, ndim = 1, mode="c"] remove_events, \
     np.ndarray[np.int32_t, ndim = 1, mode="c"] I, \
     np.ndarray[np.int32_t, ndim = 1, mode="c"] J, \
-    np.ndarray[np.int32_t, ndim = 1, mode="c"] distances, \
+    np.ndarray[np.float64_t, ndim = 1, mode="c"] distances, \
     np.ndarray[np.float64_t, ndim = 1, mode="c"] offset_visibility, \
     np.ndarray[np.float64_t, ndim = 1, mode="c"] visibility, \
     np.ndarray[np.float64_t, ndim = 2, mode="c"] visibility_map):
